Terminado

+++ src/app/components/canciones/canciones-form/canciones-form.component.ts
@@ -1,7 +1,8 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { Cancion } from '../../../models/cancion';
 import { CancionesService } from '../../../providers/canciones.service';
+import { longStackSupport } from 'q';
 
 @Component({
   selector: 'app-canciones-form',
@@ -14,6 +15,9 @@ export class CancionesFormComponent implements OnInit {
   @Input('cancion') cancion: Cancion;
   @Input('canciones') canciones: Cancion[];
   @Input('crearbtn') crearbtn: boolean;
+
+  @Output() done = new EventEmitter;
+
   formulario: FormGroup;
 
   constructor(private fb: FormBuilder, private cancionesService: CancionesService) {
@@ -30,30 +34,58 @@ export class CancionesFormComponent implements OnInit {
   }
 
   modificar(cancion: Cancion) {
-    this.cancion.nombre = this.cancion.nombre = this.formulario.controls.nombre.value;
-    this.cancionesService.putCancion(cancion).subscribe(data => {
-      this.cancion = data;
-    });
+
+    console.log(cancion);
+    
+    if (this.formulario.controls.nombre.value != '') {
+      this.cancion.nombre = this.formulario.controls.nombre.value;
+      this.cancion.nombre = this.cancion.nombre.trim();
+      this.cancionesService.putCancion(cancion).subscribe(data => {
+        this.cancion = data;
+        this.doneEvent(event);
+      }, error => {
+        console.warn(`Error al modificar ${error}`);
+      });
+    } else {
+      console.warn('El nombre no puede estar vacio');
+    }
   }
 
   eliminar(id: number) {
-    this.cancionesService.deleteCancion(id).subscribe(data => {
-      this.cancionesService.getCanciones().subscribe(data => {
-        this.canciones = data;
+    if (confirm('Realmente desea eliminar la cancion?')) {
+      this.cancionesService.deleteCancion(id).subscribe(data => {
+        this.cancionesService.getCanciones().subscribe(data => {
+          this.limpiarForm();
+          this.doneEvent(event);
+        }, error => {
+          console.warn(`Error al eliminar ${error}`);
+        });
       });
-    });
+    }
   }
 
   submit(e) {
     this.cancion = new Cancion();
+    console.log('nueva cancion %o', this.cancion);
+    
     this.cancion.nombre = this.formulario.controls.nombre.value;
-    this.cancionesService.postCancion(this.cancion).subscribe(data => {
-      this.cancionesService.getCanciones().subscribe(data => {
-        this.canciones = data;
-        console.log(this.canciones);
-        
+    this.cancion.nombre = this.cancion.nombre.trim();
+    if (this.cancion.nombre != '') {
+      this.cancionesService.postCancion(this.cancion).subscribe(data => {
+        this.cancionesService.getCanciones().subscribe(data => {
+          this.canciones = data;
+          this.limpiarForm();
+          this.doneEvent(event);
+        }, error => {
+          console.warn('Error al crear %o', error);
+        });
       });
-    });
+    } else {
+      console.warn('El nombre no puede estar vacio');
+    }
+  }
 
+  doneEvent(e) {
+    this.done.emit(e);
   }
 }
